Use more supported SQLite library

+++ RomVaultX/DB/DBSqlite.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.ComponentModel;
 using System.Data.Common;
-using System.Data.SQLite;
+using Microsoft.Data.Sqlite;
 using RomVaultX.Util;
 using RVIO;
 
@@ -10,19 +10,19 @@ namespace RomVaultX.DB
     public class DBSqlite
     {
         private const int DBVersion = 7;
-        private string _dbFilename;
-        public SQLiteConnection Connection;
+        private string? _dbFilename;
+        public SqliteConnection? Connection;
 
-        public string ConnectToDB()
+        public string? ConnectToDB()
         {
             _dbFilename = AppSettings.ReadSetting("DBFileName");
             if (_dbFilename == null)
             {
                 AppSettings.AddUpdateAppSettings("DBFileName", "rom");
-
                 _dbFilename = AppSettings.ReadSetting("DBFileName");
             }
-            string dbMemCacheSize = AppSettings.ReadSetting("DBMemCacheSize");
+
+            string? dbMemCacheSize = AppSettings.ReadSetting("DBMemCacheSize");
             if (dbMemCacheSize == null)
             {
                 // I use 8000000
@@ -34,7 +34,7 @@ namespace RomVaultX.DB
 
             bool datFound = File.Exists(_dbFilename);
 
-            Connection = new SQLiteConnection(@"data source=" + _dbFilename + ";Version=3");
+            Connection = new SqliteConnection($"data source={_dbFilename};Version=3");
             Connection.Open();
 
             // ExecuteNonQuery("PRAGMA temp_store = MEMORY");
@@ -45,18 +45,18 @@ namespace RomVaultX.DB
             ExecuteNonQuery("PRAGMA threads = 7");
             ExecuteNonQuery("PRAGMA auto_vacuum = FULL"); // Experimental pragma to reduce size of the DB file
 
-            string dbCheckOnStartup = AppSettings.ReadSetting("DBCheckOnStartup");
+            string? dbCheckOnStartup = AppSettings.ReadSetting("DBCheckOnStartup");
             if (dbCheckOnStartup == null)
             {
                 AppSettings.AddUpdateAppSettings("DBCheckOnStartup", "false");
                 dbCheckOnStartup = AppSettings.ReadSetting("DBCheckOnStartup");
             }
 
-            if (dbCheckOnStartup.ToLower() == "true")
+            if (dbCheckOnStartup.ToLowerInvariant() == "true")
             {
-                DbCommand dbCheck = new SQLiteCommand(@"PRAGMA quick_check;", Connection);
-                object res = dbCheck.ExecuteScalar();
-                string sRes = res.ToString();
+                DbCommand dbCheck = new SqliteCommand(@"PRAGMA quick_check;", Connection);
+                var res = dbCheck.ExecuteScalar();
+                string? sRes = res?.ToString();
 
                 if (sRes != "ok")
                     return sRes;
@@ -68,7 +68,7 @@ namespace RomVaultX.DB
 
             MakeTriggers();
 
-            string skipIndexing = AppSettings.ReadSetting("SkipIndexingOnStartup");
+            string? skipIndexing = AppSettings.ReadSetting("SkipIndexingOnStartup");
             if (skipIndexing != "true")
                 MakeIndex();
 
@@ -78,47 +78,40 @@ namespace RomVaultX.DB
         private void CheckDbVersion(ref bool datFound)
         {
             if (!datFound)
-            {
                 return;
-            }
 
             int testVersion = 0;
             try
             {
-                DbCommand dbVersionCommand = new SQLiteCommand(@"SELECT version from version limit 1", Connection);
-                object res = dbVersionCommand.ExecuteScalar();
+                DbCommand dbVersionCommand = new SqliteCommand(@"SELECT version from version limit 1", Connection);
+                var res = dbVersionCommand.ExecuteScalar();
 
                 if (res != null && res != DBNull.Value)
-                {
                     testVersion = System.Convert.ToInt32(res);
-                }
 
                 if (testVersion == DBVersion)
-                {
                     return;
-                }
             }
-            catch (Exception)
-            {
-            }
+            catch { }
 
-            Connection.Close();
-            File.Delete(_dbFilename);
-            Connection.Open();
+            Connection?.Close();
+            if (_dbFilename != null && File.Exists(_dbFilename))
+                File.Delete(_dbFilename);
+
+            Connection?.Open();
             datFound = false;
         }
 
         public void ExecuteNonQuery(string query, params object[] args)
         {
-            using (SQLiteCommand command = new SQLiteCommand(query, Connection))
-            {
-                for (int i = 0; i < args.Length; i += 2)
-                {
-                    command.Parameters.Add(new SQLiteParameter(args[i].ToString(), args[i + 1]));
-                }
+            using var command = new SqliteCommand(query, Connection);
 
-                command.ExecuteNonQuery();
+            for (int i = 0; i < args.Length; i += 2)
+            {
+                command.Parameters.Add(new SqliteParameter(args[i].ToString(), args[i + 1]));
             }
+
+            command.ExecuteNonQuery();
         }
 
         private void MakeDB()
@@ -176,7 +169,7 @@ namespace RomVaultX.DB
                     UPDATE GAME SET
                         RomTotal = RomTotal + 1,
                         RomGot = RomGot + (IFNULL(New.FileId,0)>0),
-                        RomNoDump = RomNoDump + (IFNULL(New.status ='nodump' and New.FileId is null,0)),
+                        RomNoDump = RomNoDump + (IFNULL(New.status ='nodump' and New.FileId IS null,0)),
                         ZipFileLength=null,
                         LastWriteTime=null,
                         CreationTime=null,
@@ -198,7 +191,7 @@ namespace RomVaultX.DB
                     UPDATE GAME SET
                         RomTotal = RomTotal - 1,
                         RomGot = RomGot - (IFNULL(Old.FileId,0)>0),
-                        RomNoDump = RomNoDump - (IFNULL(Old.status ='nodump' and Old.FileId is null,0)),
+                        RomNoDump = RomNoDump - (IFNULL(Old.status ='nodump' and Old.FileId IS null,0)),
                         ZipFileLength=null,
                         LastWriteTime=null,
                         CreationTime=null,
@@ -219,7 +212,7 @@ namespace RomVaultX.DB
                 BEGIN 
                     UPDATE GAME SET
                         RomGot = RomGot - (IFNULL(Old.FileId,0)>0) + (IFNULL(New.FileId,0)>0),
-                        RomNoDump = RomNoDump - (IFNULL(New.status ='nodump',0) and Old.FileId is null) + (IFNULL(New.status ='nodump',0) and New.FileId is null),
+                        RomNoDump = RomNoDump - (IFNULL(New.status ='nodump',0) and Old.FileId IS null) + (IFNULL(New.status ='nodump',0) and New.FileId IS null),
                         ZipFileLength=null,
                         LastWriteTime=null,
                         CreationTime=null,
@@ -280,150 +273,95 @@ namespace RomVaultX.DB
             ");
         }
 
-        public void MakeIndex(BackgroundWorker bgw = null)
+        public void MakeIndex(BackgroundWorker? bgw = null)
         {
             if (bgw == null)
-            {
                 ConsoleManager.Show();
-            }
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwRange2Visible(true));
-                bgw.ReportProgress(0, new bgwSetRange2(15));
-            }
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(0));
-                bgw.ReportProgress(0, new bgwText2("Creating Index ROM-SHA1"));
-            }
+            bgw?.ReportProgress(0, new bgwRange2Visible(true));
+            bgw?.ReportProgress(0, new bgwSetRange2(15));
+
+            bgw?.ReportProgress(0, new bgwValue2(0));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index ROM-SHA1"));
             Console.WriteLine("Creating Index 1/15: ROM-SHA1");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [ROMSHA1Index]   ON [ROM]   ([sha1]        ASC);");
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(1));
-                bgw.ReportProgress(0, new bgwText2("Creating Index ROM-MD5"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(1));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index ROM-MD5"));
             Console.WriteLine("Creating Index 2/15: ROM-MD5");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [ROMMD5Index]    ON [ROM]   ([md5]         ASC); ");
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(2));
-                bgw.ReportProgress(0, new bgwText2("Creating Index ROM-CRC"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(2));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index ROM-CRC"));
             Console.WriteLine("Creating Index 3/15: ROM-CRC");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [ROMCRCIndex]    ON [ROM]   ([crc]         ASC); ");
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(3));
-                bgw.ReportProgress(0, new bgwText2("Creating Index ROM-Size"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(3));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index ROM-Size"));
             Console.WriteLine("Creating Index 4/15: ROM-Size");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [ROMSizeIndex]   ON [ROM]   ([size]        ASC); ");
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(4));
-                bgw.ReportProgress(0, new bgwText2("Creating Index ROM-FileId"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(4));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index ROM-FileId"));
             Console.WriteLine("Creating Index 5/15: ROM-FileId");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [ROMFileIdIndex] ON [ROM]   ([FileId]      ASC); ");
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(5));
-                bgw.ReportProgress(0, new bgwText2("Creating Index ROM-GameId-Name"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(5));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index ROM-GameId-Name"));
             Console.WriteLine("Creating Index 6/15: ROM-GameId-Name");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [ROMGameId]      ON [ROM]   ([GameId]      ASC,[name] ASC);");
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(6));
-                bgw.ReportProgress(0, new bgwText2("Creating Index Game-DatId"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(6));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index Game-DatId"));
             Console.WriteLine("Creating Index 7/15: Game-DatId");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [GameDatId]      ON [GAME]  ([DatId]       ASC,[name] ASC);");
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(7));
-                bgw.ReportProgress(0, new bgwText2("Creating Index Game-DirId"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(7));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index Game-DirId"));
             Console.WriteLine("Creating Index 8/15: Game-DirId");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [GameDirId]      ON [GAME]  ([DirId]       ASC,[ZipFileLength] ASC,[name] ASC);");
 
-
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(8));
-                bgw.ReportProgress(0, new bgwText2("Creating Index FILE-SHA1"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(8));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index FILE-SHA1"));
             Console.WriteLine("Creating Index 9/15: FILE-SHA1");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [FILESHA1]       ON [FILES] ([sha1]        ASC);");
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(9));
-                bgw.ReportProgress(0, new bgwText2("Creating Index FILE-MD5"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(9));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index FILE-MD5"));
             Console.WriteLine("Creating Index 10/15: FILE-MD5");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [FILEMD5]        ON [FILES] ([md5]         ASC);");
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(10));
-                bgw.ReportProgress(0, new bgwText2("Creating Index FILE-CRC"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(10));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index FILE-CRC"));
             Console.WriteLine("Creating Index 11/15: FILE-CRC");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [FILECRC]        ON [FILES] ([crc]         ASC);");
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(11));
-                bgw.ReportProgress(0, new bgwText2("Creating Index FILE-AltSHA1"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(11));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index FILE-AltSHA1"));
             Console.WriteLine("Creating Index 12/15: FILE-AltSHA1");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [FILEAltSHA1]    ON [FILES] ([altsha1]     ASC);");
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(12));
-                bgw.ReportProgress(0, new bgwText2("Creating Index DAT-DirId"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(12));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index DAT-DirId"));
             Console.WriteLine("Creating Index 13/15: DAT-DirId");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [DATDIRID]       ON [DAT]   ([DirId]       ASC);");
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(13));
-                bgw.ReportProgress(0, new bgwText2("Creating Index Dir-ParentDirId"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(13));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index Dir-ParentDirId"));
             Console.WriteLine("Creating Index 14/15: Dir-ParentDirId");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [DIRPARENTDIRID] ON [DIR]   ([ParentDirId] ASC);");
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(14));
-                bgw.ReportProgress(0, new bgwText2("Creating Index Dir-FullName"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(14));
+            bgw?.ReportProgress(0, new bgwText2("Creating Index Dir-FullName"));
             Console.WriteLine("Creating Index 15/15: Dir-FullName");
             ExecuteNonQuery(@"CREATE INDEX IF NOT EXISTS [DIRFULLNAME]    ON [DIR]   ([fullname]    ASC);");
 
-            if (bgw != null)
-            {
-                bgw.ReportProgress(0, new bgwValue2(15));
-                bgw.ReportProgress(0, new bgwText2("Indexing Complete"));
-            }
+            bgw?.ReportProgress(0, new bgwValue2(15));
+            bgw?.ReportProgress(0, new bgwText2("Indexing Complete"));
             Console.WriteLine("Indexing Complete");
 
             if (bgw == null)
-            {
                 ConsoleManager.Hide();
-            }
         }
 
         public void DropIndex()
@@ -449,12 +387,12 @@ namespace RomVaultX.DB
 
         public DbCommand Command(string command)
         {
-            return new SQLiteCommand(command, Connection);
+            return new SqliteCommand(command, Connection);
         }
 
         public DbParameter Parameter(string param, object value)
         {
-            return new SQLiteParameter(param, value);
+            return new SqliteParameter(param, value);
         }
     }
 }
